requirements.txt
alembic==1.5.4
arena-api @ file:///home/chris/Desktop/lucid/arena_api_python/arena_api-2.0.1-py3-none-any.whl
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.8.2
bidict==0.21.2
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
dominate==2.6.0
easydict==1.9
eventlet==0.30.1
filterpy==1.4.5
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.4.4
gast==0.2.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0

idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.3

imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1

Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Mako==1.1.4

MarkupSafe==1.1.1
matplotlib==3.3.3
moviepy==1.0.1
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.5
Pillow==7.0.0
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scipy==1.4.1

sockets==1.0.0
soupsieve==2.1
SQLAlchemy==1.3.23
termcolor==1.1.0
tqdm==4.60.0
urllib3==1.25.11
visitor==0.1.3
watchdog==1.0.2
webencodings==0.5.1
websockets==8.1

wrapt==1.12.1

